requirements.txt
# StarCraft 2 Bot Dependencies
# Core SC2 API library
burnysc2>=1.0.0

# Numerical operations
numpy>=1.20.0

# PyTorch for neural network features (optional but recommended)
torch>=1.9.0

# Enhanced logging (used by bot)
loguru>=0.6.0

# Replay analysis for imitation learning
sc2reader>=1.0.0

# HTTP requests for API deployment
requests>=2.25.0

# Environment variable management (.env file support)
python-dotenv>=0.19.0

# Gemini API for self-healing system
google-genai>=0.2.0

# Phone dashboard and Google Tasks integration (optional)
flask>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Mobile dashboard backend
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
rich>=13.0.0

